--- src/components/Header/header.scss
@@ -1,28 +1,0 @@
-header {
-  // box-shadow: -2px 2px 13px -7px;
-  // border-bottom: 1px solid #e8e8e8;
-  margin-bottom: 30px;
-  padding: 5px;
-  display: flex;
-  align-items: center;
-  justify-content: space-evenly;
-  background-color: #7f8fa6;
-  .dillcap {
-    font-family: "Pacifico";
-    color: white;
-    // text-shadow: 0px 1px 11px #acacac;
-    font-size: 22px;
-  }
-}
-
-#app {
-  min-height: 97vh;
-}
-
-footer {
-  width: 100%;
-  font-size: 10px;
-  text-align: center;
-  color: #7f8fa6;
-  margin-bottom: 20px;
-}
