--- src/components/SubmissionForm/SubmissionForm.scss
@@ -1,3 +1,0 @@
-form {
-  color: teal;
-}
