--- src/components/SubmissionForm/SubmissionFormSection.scss
@@ -1,3 +1,0 @@
-section {
-  margin-bottom: 1em;
-}
