--- src/components/SubmissionForm/SubmissionFormSelect.scss
@@ -1,4 +1,0 @@
-label {
-  display: flex;
-  justify-content: space-between;
-}
