--- src/components/SubmissionForm/__tests__/CreateSubmission.test.js
@@ -1,16 +1,0 @@
-import React from 'react';
-
-import App from '../App';
-
-describe('App component', () => {
-  let wrapper;
-
-  beforeEach(() => {
-    wrapper = render(<App />);
-  });
-
-  it('renders', () => {
-    const { asFragment } = wrapper;
-    expect(asFragment()).toMatchSnapshot();
-  });
-});
