+++ src/pages/Season/styles/index.scss
