+++ src/pages/SubmissionNew/styles/index.scss
